Add even more fixes
- Fix when attacker/ killer quickly presses Q key (`lastinv` cmd.) and CSX tells the plugins a different weapon was used to cause damage/ to kill (as `Damage` msg. is sent in `UpdateClientData()` in CS/ CZ case).
- Fix when attacker disconnects/ gets kicked during the attack and CSX tells plugins that the victim committed suicide rather than being killed by a different player on the server.

+++ modules/cstrike/csx/CRank.cpp
@@ -65,7 +65,7 @@ RankSystem::RankStats::~RankStats() {
 
 void RankSystem::RankStats::setName( const char* nn  )	{
 	delete[] name;
-	namelen = strlen(nn) + 1;
+	namelen = short(strlen(nn) + 1);
 	name = new char[namelen];
 	if ( name )
 		strcpy( name , nn );
@@ -75,7 +75,7 @@ void RankSystem::RankStats::setName( const char* nn  )	{
 
 void RankSystem::RankStats::setUnique( const char* nn  )	{
 	delete[] unique;
-	uniquelen = strlen(nn) + 1;
+	uniquelen = short(strlen(nn) + 1);
 	unique = new char[uniquelen];
 	if ( unique )
 		strcpy( unique , nn );
